mcp-ionet-workspace/requirements.txt
fastmcp>=0.4.1
google-api-python-client>=2.100.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
requests>=2.31.0

